backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
httpx
translate-toolkit
rapidfuzz
motor
sqlalchemy[asyncio]
aiosqlite
asyncpg
pydantic-settings
python-dotenv
openai
anthropic
pytest
pytest-asyncio
